fix NPE, fix extra negation for bot check, reorganized code
