--- src/main/kotlin/com/lmo/ninie/io/PingController.kt
@@ -1,12 +1,0 @@
-package com.lmo.ninie.io
-
-import org.springframework.http.ResponseEntity
-import org.springframework.stereotype.Controller
-import org.springframework.web.bind.annotation.GetMapping
-
-@Controller
-class PingController {
-
-    @GetMapping(name = "/")
-    fun ping(): ResponseEntity<Nothing> = ResponseEntity.accepted().build()
-}
