--- src/main/kotlin/com/lmo/ninie/io/commands/CommandBase.kt
@@ -1,5 +1,0 @@
-package com.lmo.ninie.io.commands
-
-import com.lmo.ninie.io.services.NinieRespondable
-
-interface CommandBase<T> : NinieRespondable<T>
