--- src/main/kotlin/com/lmo/ninie/io/commands/impl/Meow.kt
@@ -1,12 +1,0 @@
-package com.lmo.ninie.io.commands.impl
-
-import com.lmo.ninie.io.commands.MessageCommandBase
-import com.lmo.ninie.io.constants.text.Emojis.HEART
-import discord4j.core.`object`.entity.Message
-import org.springframework.stereotype.Component
-
-@Component
-class Meow : MessageCommandBase {
-
-    override fun response(message: Message) = "MEEEEOOOOOOW $HEART"
-}
