+++ src/main/kotlin/com/lmo/ninie/io/extensions/eventmessage/IsBotAuthor.kt
@@ -2,4 +2,4 @@ package com.lmo.ninie.io.extensions.eventmessage
 
 import discord4j.core.`object`.entity.Message
 
-fun Message.isBotAuthor() =  this.author.map { !it.isBot }.orElse(false)
+fun Message.isBotAuthor() =  this.author.map { it.isBot }.orElse(false)
