+++ src/main/kotlin/com/lmo/ninie/io/interactions/commands/CommandBase.kt
@@ -1,0 +1,5 @@
+package com.lmo.ninie.io.interactions.commands
+
+import com.lmo.ninie.io.interactions.NinieRespondable
+
+interface CommandBase<T> : NinieRespondable<T>
