+++ src/main/kotlin/com/lmo/ninie/io/interactions/commands/impl/Meow.kt
@@ -1,0 +1,12 @@
+package com.lmo.ninie.io.interactions.commands.impl
+
+import com.lmo.ninie.io.interactions.commands.MessageCommandBase
+import com.lmo.ninie.io.constants.text.Emojis.HEART
+import discord4j.core.`object`.entity.Message
+import org.springframework.stereotype.Component
+
+@Component
+class Meow : MessageCommandBase {
+
+    override fun response(message: Message) = "MEEEEOOOOOOW $HEART"
+}
