+++ src/main/kotlin/com/lmo/ninie/io/services/commands/impl/AliasServiceImpl.kt
@@ -1,8 +1,8 @@
 package com.lmo.ninie.io.services.commands.impl
 
-import com.lmo.ninie.io.commands.Alias
-import com.lmo.ninie.io.commands.CommandBase
-import com.lmo.ninie.io.commands.impl.*
+import com.lmo.ninie.io.interactions.commands.Alias
+import com.lmo.ninie.io.interactions.commands.CommandBase
+import com.lmo.ninie.io.interactions.commands.impl.*
 import com.lmo.ninie.io.services.commands.AliasService
 import discord4j.discordjson.json.MessageData
 import io.vavr.kotlin.option
@@ -20,22 +20,22 @@ class AliasServiceImpl(
     val unknown: Unknown
 ) : AliasService {
 
-    override fun all(): List<Alias> = Alias::class.nestedClasses.map { it.objectInstance as Alias }
+    override fun all(): List<Alias> = Alias.values().toList()
 
     override fun find(alias: String?) = all()
         .find { e -> e.defaultAlias == alias }
         .option()
 
-    override fun mapToCommand(alias: Alias): CommandBase<MessageData> = when (alias) {
-        is Alias.About -> about
-        is Alias.Cry -> cry
-        is Alias.Fart -> fart
-        is Alias.Help -> help
-        is Alias.Meow -> meow
-        is Alias.Sing -> sing
-        is Alias.Songs -> songs
-        is Alias.Unknown -> unknown
-    }
-
+    override fun mapToCommand(alias: Alias): CommandBase<MessageData> =
+        when (alias) {
+            Alias.About -> about
+            Alias.Cry -> cry
+            Alias.Fart -> fart
+            Alias.Help -> help
+            Alias.Meow -> meow
+            Alias.Sing -> sing
+            Alias.Songs -> songs
+            Alias.Unknown -> unknown
+        }
 
 }
