+++ src/main/kotlin/com/lmo/ninie/io/services/dao/SongService.kt
@@ -1,0 +1,12 @@
+package com.lmo.ninie.io.services.dao
+
+import com.lmo.ninie.io.models.songs.Song
+import io.vavr.control.Option
+
+interface SongService {
+    fun any(): Song
+
+    fun find(name: String): Option<Song>
+
+    fun all(): List<Song>
+}
