--- src/main/kotlin/com/lmo/ninie/io/services/reactions/RespondableMapperService.kt
@@ -1,10 +1,0 @@
-package com.lmo.ninie.io.services.reactions
-
-import discord4j.core.`object`.entity.Message
-import discord4j.discordjson.json.MessageData
-import reactor.core.publisher.Mono
-
-interface RespondableMapperService {
-    fun reactToCreation(message: Message): Mono<Unit>
-    fun reactToUpdate(message: Message): Mono<Unit>
-}
